feat(): segregate into another nuget lib for netcoreapp3/netstandard
