+++ .editorconfig
@@ -15,7 +15,7 @@ indent_style = space
 
 # Indentation and spacing
 indent_size = 4
-indent_style = tab
+indent_style = space
 tab_width = 4
 
 # New line preferences
