+++ SignalR.Orleans.sln
@@ -1,6 +1,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27004.2009
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29424.173
 MinimumVisualStudioVersion = 15.0.26124.0
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{20D142F6-AE1A-4AF1-A314-13E6AAF040ED}"
 EndProject
@@ -16,6 +16,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 		README.Nuget.md = README.Nuget.md
 	EndProjectSection
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SignalR.Orleans.AspNet", "src\SignalR.Orleans.AspNet\SignalR.Orleans.AspNet.csproj", "{F244A392-59B6-40D1-BDDE-04CD27526CD0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -51,6 +53,18 @@ Global
 		{C047C695-9F3C-4DDE-9D0C-06608E191B67}.Release|x64.Build.0 = Release|Any CPU
 		{C047C695-9F3C-4DDE-9D0C-06608E191B67}.Release|x86.ActiveCfg = Release|Any CPU
 		{C047C695-9F3C-4DDE-9D0C-06608E191B67}.Release|x86.Build.0 = Release|Any CPU
+		{F244A392-59B6-40D1-BDDE-04CD27526CD0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F244A392-59B6-40D1-BDDE-04CD27526CD0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F244A392-59B6-40D1-BDDE-04CD27526CD0}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F244A392-59B6-40D1-BDDE-04CD27526CD0}.Debug|x64.Build.0 = Debug|Any CPU
+		{F244A392-59B6-40D1-BDDE-04CD27526CD0}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F244A392-59B6-40D1-BDDE-04CD27526CD0}.Debug|x86.Build.0 = Debug|Any CPU
+		{F244A392-59B6-40D1-BDDE-04CD27526CD0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F244A392-59B6-40D1-BDDE-04CD27526CD0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F244A392-59B6-40D1-BDDE-04CD27526CD0}.Release|x64.ActiveCfg = Release|Any CPU
+		{F244A392-59B6-40D1-BDDE-04CD27526CD0}.Release|x64.Build.0 = Release|Any CPU
+		{F244A392-59B6-40D1-BDDE-04CD27526CD0}.Release|x86.ActiveCfg = Release|Any CPU
+		{F244A392-59B6-40D1-BDDE-04CD27526CD0}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -58,6 +72,7 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{559ED786-70A6-4BBB-8F09-C544F8E9D7E6} = {20D142F6-AE1A-4AF1-A314-13E6AAF040ED}
 		{C047C695-9F3C-4DDE-9D0C-06608E191B67} = {DB316F2F-30C5-4B15-916F-3C1A115E2532}
+		{F244A392-59B6-40D1-BDDE-04CD27526CD0} = {20D142F6-AE1A-4AF1-A314-13E6AAF040ED}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {E55056FC-21D0-4770-A905-172EA054FF14}
