+++ src/SignalR.Orleans.AspNet/Properties/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+﻿
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("SignalR.Orleans.Tests")]
+public class AssemblyInfo
+{
+}
