+++ src/SignalR.Orleans/SignalR.Orleans.csproj
@@ -2,18 +2,17 @@
 
   <PropertyGroup>
     <PackageId>Sketch7.SignalR.Orleans</PackageId>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
+    <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
 
   <PropertyGroup>
     <Title>SignalR.Orleans - Orleans backplane for SignalR Core</Title>
     <Description>Orleans backplane for SignalR Core.</Description>
-    <PackageTags>$(PackageTags) AspNetCore</PackageTags>
+    <PackageTags>$(PackageTags)</PackageTags>
     <PackageReleaseNotes></PackageReleaseNotes>
   </PropertyGroup>
 
   <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="Microsoft.Orleans.CodeGenerator.MSBuild" Version="$(OrleansVersion)" PrivateAssets="all">
      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
    </PackageReference>
@@ -21,6 +20,7 @@
     <PackageReference Include="Microsoft.Orleans.Core" Version="$(OrleansVersion)" />
     <PackageReference Include="Microsoft.Orleans.OrleansProviders" Version="$(OrleansVersion)" />
     <PackageReference Include="Microsoft.Orleans.Runtime.Abstractions" Version="$(OrleansVersion)" />
+    <PackageReference Include="Microsoft.AspNetCore.SignalR.Common" Version="3.0.0" />
   </ItemGroup>
 
 </Project>
